python/requirements.txt
pyserial
rpi-lcd